python/requirements.txt
openai-agents==0.0.4
flask>=2.2.0
flask-cors>=3.0.10
websocket-client>=1.3.2
websockets>=11.0.3
pillow>=9.3.0
numpy>=1.24.0
python-dotenv>=0.20.0
gunicorn>=20.1.0
aiohttp>=3.8.4
requests>=2.28.2
